Third commit

+++ src/test/java/framework/PatternHelper.java
@@ -6,12 +6,12 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class PatternHelper {
-    public static int getIntFromString(String origin) {
+    public static double getIntFromString(String origin) {
 
         Pattern pattern = Pattern.compile("\\d+[.]?\\d*");
         Matcher matcher = pattern.matcher(origin);
         if (matcher.find()) {
-            return Integer.parseInt(matcher.group());
+            return Double.parseDouble(matcher.group());
         } else {
             Logger.getAnonymousLogger().log(Level.SEVERE, "Ошибка при преобразовании к числу: число не найдено");
             return 0;
